fix: use Long instead of Integer for primary keys

+++ src/main/java/com/isko_d/isko_d/exception/NotFoundException.java
@@ -2,7 +2,7 @@ package com.isko_d.isko_d.exception;
 
 public class NotFoundException extends RuntimeException {
 
-    public NotFoundException(Class<?> entity, Integer id) {
+    public NotFoundException(Class<?> entity, Long id) {
         super(String.format("%s with id %d not found.", entity.getSimpleName(), id));
     }
